backend_app/amacars_webapp/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.40
cryptography==36.0.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
Flask-CORS==4.0.0

